examples/sentiment/requirements.txt
bert-tensorflow==1.0.1
tensorflow_hub==0.5.0
